+++ dp/_dp_array.py
@@ -4,6 +4,7 @@ import warnings
 import numpy as np
 
 from dp._logger import Logger, Op
+from dp._index_converter import _nd_slice_to_indices
 
 
 class DPArray:
@@ -41,7 +42,6 @@ class DPArray:
 
         self._logger = Logger() if logger is None else logger
         self._logger.add_array(array_name, shape)
-        self._logger_enabled = True
 
         self._array_name = array_name
 
@@ -66,69 +66,6 @@ class DPArray:
 
         raise ValueError("Unsupported dtype. Must be np.float32 or"
                          "np.float64")
-
-    def _nd_slice_to_indices(self, nd_slice):
-        """Converts a nd-slice to indices.
-
-        Calculate the indices from the slices.
-        1. Get the start, stop, and step values for each slice.
-        2. Use np.arange to create arrays of indices for each slice.
-        3. Create 1D array of indices using meshgrid and column_stack.
-
-        Args:
-            nd_slice (slice/int, list/tuple of slice/int):
-                nd_slice can be anything used to index a numpy array. For
-                example:
-                - Direct indexing: (0, 1, ...) or [0, 1, ...]
-                - 1d slice: slice(0, 10, 2)
-                - nd slice: (slice(0, 10, 2), slice(1, 5, 1), ...)
-                - Mixture: (slice(0, 10, 2), 5, 1)
-
-        Returns:
-            list of tuples/integer: length n list of d-tuples, where n is the 
-                number of indices and d is the dimension DPArray. If d = 1, 
-                then the list will contain integers instead.
-
-        Raises:
-            ValueError: ``nd_slice`` is not a slice object, a list of slice
-                objects, or a list of tuples of integers.
-            ValueError: Some element in ``nd_slice`` is not a valid slice
-                object or integer.
-        """
-        if isinstance(nd_slice, (slice, int)):
-            # Convert 1d slice to nd slice.
-            nd_slice = (nd_slice,)
-        if isinstance(nd_slice, list):
-            return nd_slice
-        if not isinstance(nd_slice, (list, tuple)):
-            raise ValueError(f"'nd_slice' has type {type(nd_slice)}, must be "
-                             f"a slice object, a list/tuple of slice objects,"
-                             f" or a list/tuple of integers.")
-
-        slice_indices = []
-        for dim, size in enumerate(self._arr.shape):
-            s = nd_slice[dim]
-            if isinstance(s, slice):
-                # Handle slice objects.
-                slice_indices.append(np.arange(*s.indices(size)))
-            elif isinstance(s, int):
-                # Handle tuple of integers for direct indexing.
-                slice_indices.append(s)
-            else:
-                raise ValueError("Each element in 'nd_slice' must be a valid "
-                                 "slice object or integer.")
-
-        # Generate the meshgrid of indices and combine indices into
-        # n-dimensional index tuples.
-        mesh_indices = np.meshgrid(*slice_indices, indexing="ij")
-        indices = np.stack(mesh_indices,
-                           axis=-1).reshape(-1, len(slice_indices))
-
-        # Convert to tuple if index is > 1D, otherwise remove the last
-        # dimension.
-        indices_tuples = ([tuple(row) for row in indices] if indices.shape[1]
-                          != 1 else np.squeeze(indices, axis=1))
-        return indices_tuples
 
     def get_timesteps(self):
         """Retrieve the timesteps of all DPArrays associated with this array's 
@@ -181,9 +118,8 @@ class DPArray:
                 f"'{self._array_name}'. Undefined elements: "
                 f"{undef_read_indices}.",
                 category=RuntimeWarning)
-        log_idx = self._nd_slice_to_indices(idx)
-        if self._logger_enabled:
-            self._logger.append(self._array_name, Op.READ, log_idx)
+        log_idx = _nd_slice_to_indices(self._arr, idx)
+        self._logger.append(self._array_name, Op.READ, log_idx)
         return self._arr[idx]
 
     def __setitem__(self, idx, value):
@@ -196,7 +132,7 @@ class DPArray:
         Raises:
             ValueError: If ``idx`` is a slice object.
         """
-        log_idx = self._nd_slice_to_indices(idx)
+        log_idx = _nd_slice_to_indices(self._arr, idx)
 
         value = self.dtype(value)
         if isinstance(value, self.dtype):
@@ -204,8 +140,7 @@ class DPArray:
 
         self._arr[idx] = value.reshape(self._arr[idx].shape)
         self._occupied_arr[idx] = True
-        if self._logger_enabled:
-            self._logger.append(self._array_name, Op.WRITE, log_idx, value)
+        self._logger.append(self._array_name, Op.WRITE, log_idx, value)
 
     def __eq__(self, other):
         """Equal to operator.
@@ -246,7 +181,7 @@ class DPArray:
         elements = [0, 1, 2, 3, 4, 4]
         indices = [None, 2, 4, 6, 8, 1]
         ```
-        The output of this function will be `4` and indices `8`, `1` will be
+        The output of this function will be `4` and indices `8` and `1` will be
         highlighted.
 
         Args:
@@ -277,28 +212,27 @@ class DPArray:
         for i, e in zip(indices, elements):
             # Unravel when index is a slice.
             if isinstance(i, slice) and isinstance(e, np.ndarray):
-                # get argmax indices
+                # Get argmax indices.
                 slice_max_idx = e.flatten().argmax()
-                slice_indices = self._nd_slice_to_indices(i)
+                slice_indices = _nd_slice_to_indices(self._arr, i)
                 i = slice_indices[slice_max_idx]
-                # get max element
+                # Get max element.
                 e = np.max(e)
             else:
-                # make index into a singleton
+                # Make index into a singleton.
                 i = [i]
 
-            # If new best index/element is found
+            # If new best index/element is found.
             if cmp(e, best_element):
                 best_indices = i
                 best_element = e
 
-            # If index has equivalent element to the best element
+            # If index has equivalent element to the best element.
             elif e == best_element:
                 best_indices.extend(i)
 
         # Highlight and write value.
-        if self._logger_enabled:
-            self.logger.append(self._array_name, Op.HIGHLIGHT, best_indices)
+        self.logger.append(self._array_name, Op.HIGHLIGHT, best_indices)
         return best_element
 
     def max(self, indices, elements):
@@ -333,89 +267,20 @@ class DPArray:
         """
         return self._cmp(lambda x, y: x < y, indices, elements)
     
-    def backtrack(self, endpoint=None):
-        """
-        Get a backtracked solution to this DPArray object.
+    def add_traceback_path(self, path):
+        """Add a traceback path to this DPArray object.
+        
+        Paths added to a DPArray object will be displayed when calling display()
+        on that DPArray object. The path will appear on the last frame of the
+        visualization window (slider is in the rightmost position).
 
         Args:
-            endpoint (int or tuple): index of the endpoint of the backtracking.
-            The default value is the largest index (largest in each dimension).
-        
-        Return:
-            list: a list of indices in a backtracked solution.
-            Solutions may not be unique. This function returns an arbitrary solution.
-
-        Raises:
-            IndexError: endpoint references an unitinialized element of the DPArray
+            path (list of tuples): A list of indices to be displayed.
+            Indices should be tuples with as many elements as the dimension
+            of the array (the number of dimensions is given by len(shape)).
         """
-        # Start at largest indexed element by default
-        if endpoint is None:
-            endpoint = tuple([i - 1 for i in self.shape])
-
-        # Raise error if endpoint is uninitialized
-        if not self.occupied_arr[endpoint]:
-            raise IndexError(f'Backtracking failed because endpoint {endpoint} is unitialized')
-
-        # construct solution from timesteps
-        solution = [endpoint]
-        for timestep in reversed(self.get_timesteps()):
-            # If the node being written is the predecessor
-            if solution[-1] in timestep[self.array_name][Op.WRITE]:
-                # No predecessor so stop backtracking
-                if len(timestep[self.array_name][Op.HIGHLIGHT]) == 0:
-                    break
-                # Add an arbitrary predecessor to the solution
-                for predecessor in timestep[self.array_name][Op.HIGHLIGHT]:
-                    break
-                solution.append(predecessor)
-        
-        return solution[::-1]
-
-    def is_backtrack(self, solution):
-        """
-        Validate if a given backtrack solution is correct.
-
-        Args:
-            solution (list of indices): a list of indices to validate.
-            For 1D DPArrays, this should be a list of integers.
-            For 2D DPArrays, this should be a list of tuples.
-
-        Return:
-            Returns False if the given solution is not correct
-            and true if it is correct.
-        """
-        # Handle trivial case
-        if(len(solution) == 0):
-            return True
-        
-        # Ensure each index in the solution is initialized
-        if not np.all(self._occupied_arr[solution]):
-            return False
-
-        # Go through time steps and check predecessors
-        i = len(solution) - 1
-        for timestep in reversed(self.get_timesteps()):
-            # If writing solution[i], analyze predecessors
-            if solution[i] in timestep[self.array_name][Op.WRITE]:
-                # Check that solution[0] has no predecessors
-                if i == 0:
-                    return len(timestep[self.array_name][Op.HIGHLIGHT]) == 0
-                
-                # Check that solution[i - 1] is a predecessor of solution[i]
-                else:
-                    # solution[i - 1] is not a predecessor, so the given solution is not correct
-                    if solution[i - 1] not in timestep[self.array_name][Op.HIGHLIGHT]:
-                        return False
-                    i = i - 1
-        return True
-
-    def add_backtrack_solution(self, solution):
-        log_idx = self._nd_slice_to_indices(solution)
+        log_idx = _nd_slice_to_indices(self._arr, path)
         self._logger.append(self._array_name, Op.READ, log_idx)
-
-    def enable_logger(self, enable=True):
-        """Enable or disable logger."""
-        self._logger_enabled = enable
 
     @property
     def arr(self):