+++ HISTORY.md
@@ -3,9 +3,9 @@
 ## 0.1.0
 
 ### API
-
 - Add support for displaying multiple arrays(#23)
-- Display dependencies on click (#25)
+- Traceback solution verification and visualization (#24)
+- Display dependencies on click (#25, #26)
 - Migrate from graph objects to Dash (#20)
 - Enable arrow keys for time travel (#20)
 - Remove "null" text in empty cells (#21)